Implement admin takes points from user function
Decided to allow an admin to enter a negative number and 'take points'
from user that way. This solution workd with current code and the only
change necessary was the label on the button so that the user knows he
can award or take points using the same button.

Other minor changes:

- The instance variable inside of current_user methor in application
  controller was clashing with other instance variables. This was
causing the admin users edit page to show the incorrect user as logged
in on this page.

- The admin user's edit page was styled slightly different for it to be
  easier to read.