Implement admins can edit rewards feature
Decided to add a new admin index page for users to choose the desired
rewards they wish to edit. Before, admins had the ability to add a
rewards from the rewards index that regular users have access to.
Instead of adding too much logic to that page in which we would show
edit/delete buttons to admins and not to regular users I decided to
simply have a rewards index page that is only accessible to admins
(admin rewards index). This will also make it simple to add new
features only available to admins.